requirements.txt
PySide6
cx_Freeze
PyYAML
